Connected input data to API calls
The form inputs were not being used to call data - made that work
Consolidated one API call back into the portal (as intended)

Have not yet implemented the throttler. There are some response headers
indicating that there may be a new higher limit than the documentation
said. We should re-read the documentation, as it probably has been or
will be changed.